requirements.txt
assemblyai==0.42.0
python-dotenv>=1.0.0
